+++ package.json
@@ -36,6 +36,7 @@
     "@tippyjs/react": "^4.1.0",
     "@types/next-auth": "^3.15.0",
     "add": "^2.0.6",
+    "antd": "^4.18.6",
     "aws-sdk": "^2.819.0",
     "axios": "^0.19.2",
     "classnames": "^2.3.1",
