+++ src/components/Canvas/index.tsx
@@ -1,132 +1,47 @@
-import { useWorkspace } from '@context/workspace-context';
-import React, { useRef, useState } from 'react';
-import { PanZoom } from 'react-easy-panzoom';
-import { useSelection } from '~/src/context/selection-context';
+import { useEffect, useRef, useState } from 'react';
+import { Graph, Node } from '@antv/x6';
 import withApollo from '~/src/hooks/withApollo';
-import DraggableBlock from '../Block';
+
+const data = [
+  {
+    id: '1',
+    width: 100,
+    height: 100,
+    shape: 'rect',
+    label: 'data',
+    x: 100,
+    y: 100,
+  },
+];
 
 export default withApollo((props) => {
-  const { workspaceOptions, setWorkspaceOptions, zoomScale, setZoomScale } =
-    useWorkspace();
+  const container = useRef();
+  const [graph, setGraph] = useState(null);
 
-  const [canvasSpace, setCanvasSpace] = useState({
-    scrollAreaHeight: null,
-    scrollAreaWidth: null,
-    maxHeight: null,
-    maxWidth: null,
-    timer: null,
-    workspaceOptions,
-    setWorkspaceOptions,
-  });
+  // React.useEffect(() => {
+  //   if (root.current) {
+  //     const g = new Graph({
+  //       container: root.current,
+  //     });
+  //     g.scale(1.5);
+  //     g.fromJSON(data);
+  //     setGraph(g);
+  //   }
+  // }, [root]);
 
-  const [dragStates, setDragStates] = useState({
-    maxZIndex: 1000,
-  });
+  useEffect(() => {
+    if (container.current) {
+      const graph = new Graph({
+        container: container.current,
+        width: 800,
+        height: 300,
+        grid: true,
+      });
 
-  const panZoomRef = useRef(null);
-
-  const { selectedConnection, setSelectedConnection, canvasBlocks } =
-    useSelection();
-
-  const preventPan = (event, x, y) => {
-    if (
-      event.target &&
-      event.target.parentElement &&
-      event.target.parentElement.className &&
-      event.target.parentElement.className.includes(
-        'block' ||
-          'draggable' ||
-          'react-draggable' ||
-          'header' ||
-          'nested-canvas' ||
-          'draggable-block-container--expanded' ||
-          'draggable-block-container' ||
-          'block',
-      )
-    ) {
-      return true;
+      graph.fromJSON(data);
+      setGraph(graph);
     }
-  };
+  }, [container]);
 
-  const onZoomEnd = (e) => {
-    if (panZoomRef.current) {
-      setZoomScale(panZoomRef.current.scale);
-    }
-  };
-
-  let dragging = false;
-
-  if (!canvasBlocks && canvasBlocks.length === 0) {
-    return (
-      <div className={`loading-screen fullscreen`}>
-        <p
-          style={{
-            marginBottom: 20,
-          }}
-        >
-          You can blocks to your canvas using the Grid View.
-        </p>
-        <Button onClick={_switchToGridFormation} icon={Formations.GRID.icon}>
-          Switch to Grid View
-        </Button>
-      </div>
-    );
-  }
-
-  return (
-    <div className="canvas-container">
-      {/* <ZoomTools panZoomRef={panZoomRef} /> */}
-      <PanZoom
-        ref={panZoomRef}
-        preventPan={preventPan}
-        className="canvas"
-        onZoomEnd={onZoomEnd}
-        scale={zoomScale}
-        minZoom={0.4}
-        autoCenter={true}
-        style={{
-          width: '100vw',
-          height: '100vh',
-          WebkitFilter: 'blur(0)',
-          willChange: 'unset',
-          position: 'relative',
-        }}
-        maxZoom={3}
-        disableScrollZoom
-        onPanStart={(e) => {
-          dragging = false;
-        }}
-        onPan={() => {
-          dragging = true;
-        }}
-        onPanEnd={(e) => {
-          if (
-            !(e.target && e.target.className) &&
-            e.target.className.includes('block') &&
-            !dragging
-          ) {
-            setSelectedConnection(null);
-          }
-
-          dragging = false;
-        }}
-      >
-        {canvasBlocks.map((block, i) => (
-          <DraggableBlock
-            title={block.title ? block.title : null}
-            type={block.__typename}
-            dragStates={dragStates}
-            setDragStates={setDragStates}
-            panZoomRef={props.panZoomRef}
-            key={block.id}
-            block={block}
-            canvasSpace={canvasSpace}
-            setCanvasSpace={setCanvasSpace}
-            bounds="parent"
-            {...props}
-          />
-        ))}
-      </PanZoom>
-    </div>
-  );
+  return <div id="container" ref={container}></div>;
 });
